+++ src/test/java/ru/deliveryClub/FirstTest.java
@@ -6,15 +6,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.chrome.ChromeDriver;
 
-public class FirstTest {
-
-    public ChromeDriver driver;
-    @Before
-    public void setUp() {
-        System.setProperty("webdriver.chrome.driver", "d://Users//user//Downloads//chromedriver.exe");
-        driver = new ChromeDriver();
-        System.out.println("Test start");
-    }
+public class FirstTest extends WebDriverSettings {
 
     @Test
     public void firstTest() {
@@ -34,9 +26,4 @@ public class FirstTest {
         Assert.assertTrue(title.equals("Доставка еды из ресторанов Нижнего Новгорода от 45 минут! Delivery Club"));
     }
 
-    @After
-    public void Close() {
-        System.out.println("Test close");
-        driver.quit();
-    }
 }
