+++ src/test/java/ru/deliveryClub/WebDriverSettings.java
@@ -1,0 +1,21 @@
+package ru.deliveryClub;
+
+import org.junit.After;
+import org.junit.Before;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+public class WebDriverSettings {
+
+    public ChromeDriver driver;
+    @Before
+    public void setUp() {
+        System.setProperty("webdriver.chrome.driver", "d://Users//user//Downloads//chromedriver.exe");
+        driver = new ChromeDriver();
+    }
+
+    @After
+    public void Close() {
+        driver.quit();
+    }
+
+}
